解决sdk指令无法使用（由于运行sdk会切换到env/tools，但是bsp_root没有跟着改变），同时将get_rtt_root函数移到cm_menuconfig.py，每次运行menuconfig的时候重新从环境变量或Kconfig中获取RTT_ROOT
